requirements.txt
msal>=1.28.0
requests>=2.31.0
PyYAML>=6.0.1
rich>=13.7.0
